Redone Movement and Collision Logic 💔 ❌
In an attempt to make the collisions of the platforms , the player movement was changes to standard WASD because i believed the thrusting was causing glitching through the platforms. This remains a persistent glitch. Even with WASD before this i also tried moving all the platforms to be something  static which also didn't work. I've since tried different variations of SimulateMovement code to figure out the the collisions of the platforms each one causes glitching